Adds Tests for coverage, supporting for working with targets and a fix for a bug that was found.
