+++ SHA512SUM.txt
@@ -1,2 +1,2 @@
-9c5aca2b0b027235a9cb3f27092d84f9760f5c2a203f529110009ae9c8a1c0a92da53bacd3b80ee6251e0c61ed0b349c1fa068d11402ff2e1eaab314ce982918  Jagati.cmake
+048e41af2ae39c5cb504a5b41b453471f0081c67fc8c7a54f05ff68a2a6054be97a6505e0c4bef66e29f44ca944ceee280562ab001e6377ab61d0f957a81f4b5  Jagati.cmake
 cf83e1357eefb8bdf1542850d66d8007d620e4050b5715dc83f4a921d36ce9ce47d0d13c5d85f2b0ff8318d2877eec2f63b931bd47417a81a538327af927da3e  Index.cmake
